+++ packages/kmp-compose/src/commonMain/kotlin/ly/neptune/signal/Signal.kt
@@ -1,0 +1,8 @@
+package ly.neptune.signal
+
+object Signal {
+    const val StandardVersion = "0.1.0"
+    const val SdkVersion = "0.1.0"
+    const val Name = "Neptune. Signal"
+}
+