+++ packages/kmp-compose/src/commonMain/kotlin/ly/neptune/signal/theme/SignalSpacing.kt
@@ -1,0 +1,29 @@
+package ly.neptune.signal.theme
+
+import androidx.compose.ui.unit.dp
+
+object SignalSpacing {
+    val x1 = 4.dp
+    val x2 = 8.dp
+    val x3 = 12.dp
+    val x4 = 16.dp
+    val x5 = 20.dp
+    val x6 = 24.dp
+    val x8 = 32.dp
+    val x10 = 40.dp
+}
+
+object SignalRadius {
+    val xs = 8.dp
+    val sm = 12.dp
+    val md = 16.dp
+    val lg = 22.dp
+    val xl = 28.dp
+}
+
+object SignalSize {
+    val touchTarget = 44.dp
+    val iconButton = 44.dp
+    val navHeight = 88.dp
+}
+